+++ CHANGELOG.md
@@ -17,6 +17,10 @@ For a full diff see [`2.9.0...main`][2.9.0...main].
 * Required only implementations of `Config\RuleSet\ExplicitRuleSet` not to configure any rules for rule sets ([#313]), by [@localheinz]
 * Required implementations of `Config\RuleSet\ExplicitRuleSet` to configure non-deprecated rules that are configurable with an explicit configuration when enabled ([#314]), by [@localheinz]
 
+### Fixed
+
+* Stopped configuring rules using deprecated configuration options ([#319]), by [@localheinz]
+
 ## [`2.9.0`][2.9.0]
 
 For a full diff see [`2.8.0...2.9.0`][2.8.0...2.9.0].
@@ -306,6 +310,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#311]: https://github.com/ergebnis/php-cs-fixer-config/pull/311
 [#313]: https://github.com/ergebnis/php-cs-fixer-config/pull/313
 [#314]: https://github.com/ergebnis/php-cs-fixer-config/pull/314
+[#319]: https://github.com/ergebnis/php-cs-fixer-config/pull/319
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler